# src/main/resources/api/api-spec.yaml
openapi: 3.0.3
info:
  title: Aura Produto API
  description: API do Microsserviço de Produtos para o Sistema de Gerenciamento de Pedidos
  version: 1.0.0
  contact:
    name: Equipe Aura
    email: equipe@aura.com.br
servers:
  - url: 'http://localhost:8081'
    description: Servidor local
tags:
  - name: products
    description: Operações relacionadas a produtos

paths:
  /products:
    get:
      tags:
        - products
      summary: List all products
      operationId: listProducts
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - products
      summary: Cria um novo produto
      description: Adiciona um novo produto ao catálogo
      operationId: createProduct
      requestBody:
        description: Dados do produto a ser criado
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '201':
          description: Produto criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Dados inválidos ou SKU duplicado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /products/{sku}:
    get:
      tags:
        - products
      summary: Busca um produto pelo SKU
      description: Retorna um produto específico com base no SKU fornecido
      operationId: getProductBySku
      parameters:
        - name: sku
          in: path
          description: SKU do produto a ser recuperado
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Produto encontrado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Produto não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Identificador único do produto
        name:
          type: string
          description: Nome do produto
        sku:
          type: string
          description: Código único de identificação do produto (Stock Keeping Unit)
        price:
          type: number
          format: double
          description: Preço do produto
        description:
          type: string
          description: Descrição detalhada do produto
      required:
        - id
        - name
        - sku
        - price
        - description

    Error:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: Data e hora da ocorrência do erro
        status:
          type: integer
          format: int32
          description: Código de status HTTP
        error:
          type: string
          description: Tipo do erro
        message:
          type: string
          description: Mensagem detalhada do erro
        path:
          type: string
          description: Caminho da requisição que gerou o erro